hal/*: change base address to uint64_t to allow larger PCIe addresses
As we have changed the BAR number constants to make use of
64-bit variables, we also changed the base addresses of
our components.

+++ hal/sm_io/chips/sm_ch_24aa64.h
@@ -26,7 +26,7 @@ typedef struct _smch_24aa64_t smch_24aa64_t;
 /***************** Our methods *****************/
 
 /* Creates a new instance of the SMCH 24AA64 */
-smch_24aa64_t * smch_24aa64_new (smio_t *parent, uint32_t base, uint32_t addr,
+smch_24aa64_t * smch_24aa64_new (smio_t *parent, uint64_t base, uint32_t addr,
         int verbose);
 /* Destroy an instance of the SMCH 24AA64 */
 smch_err_e smch_24aa64_destroy (smch_24aa64_t **self_p);
